Tools: Maya cords to Panda
I'm still kinda of confused about the transition from Maya's coordinates to Panda's. This is a basic script to make that transition, however batch files apparently cannot operate decimals, so you are left to do the multiplication and division yourself. When I have more time, maybe I'll make it properly