VP Sequence, Redid VP reparenting
Aftermath of the VP Battle.  After being defeated by the Toons, the Senior VP falls from the top of Sellbot Towers and lands in the pit in the Sellbot HQ Courtyard.  He is then pushed upright with help of the headless Hollywoods.  He then does a 180 out of the pit.
Also, reparentted the VP parts.  I had trouble connecting the torso to the tank unfortunately.