Added a Health Bar for the Cog Bosses, Currently only works for VP and CJ, changed all boss models to VP
Added a Health Bar for the Cog Bosses. Works for the VP and CJ.  CFO and CEO it crashes with errors.  Changed all models to VP because im a VP Fan.  I kinda want to get rid of the decimal in the front of health since atm for example VP its 100.0/100
I want to try to get it to 100/100