InventoryBase: Fix SetInventory magic word
SetInventory originally did not restock level 7 gags when the maxLevelIndex argument was 6. The only way to restock level 7 gags using magic words was to use ~MaxToon. This has been fixed to work as intended now. The other problem was that using 6 as the maxLevelIndex argument originally did not fully restock the invoker's gags because the function was counting the level 7 gags as inventory space. (This means if the invoker had 6 gag tracks and 0 gags in their inventory, only 74 gags were restocked, along with the invoker's level 7's). This too, has been fixed so that when using 6 as the max level index, the function restocks the invoker's level 7 gags, and completely fills their inventory.